Fixed flow signature and started implementing it

+++ include/flow.h
@@ -4,7 +4,7 @@
 #include "Eigen/Dense"
 
 void flow(Eigen::MatrixXd &Vc, Eigen::MatrixXi &Ec,
-          Eigen::MatrixXd &Vf, Eigen::MatrixXd &Ef,
+          Eigen::MatrixXd &Vf, Eigen::MatrixXi &Ef,
           std::vector<Eigen::MatrixXd> &Vf_flow);
 
 #endif //CAGES_FLOW_H
